Clean recipe database
nettoyage des dernières "recettes" présentes dans la classe database

+++ controller/AppartementController.php
@@ -6,7 +6,7 @@ use function PHPSTORM_META\elementType;
  * ETML
  * Auteur : Jeremiah Steiner
  * Date: 27.04.2021
- * Controler pour gérer les recettes
+ * Controler pour gérer les appartements
  */
 
 //include_once 'model/RecetteRepository.php';
@@ -98,7 +98,7 @@ class AppartementController extends Controller {
         $database = new Database();
 
         $startIndex = 0;
-        $lengthAppartement = 5; // UTIL : modifier si on veut pouvoir modifier le nombre de recette affichée
+        $lengthAppartement = 5; // UTIL : modifier si on veut pouvoir modifier le nombre d'appartement affiché
         $_SESSION["appartementsPerPage"] = $lengthAppartement;
 
         if (array_key_exists("start", $_GET) && $_GET["start"] > 0) // si le paramettre de start n'est pas négatif
@@ -148,7 +148,7 @@ class AppartementController extends Controller {
         $database = new Database();
 
         $startIndex = 0;
-        $lengthAppartement = 5; // UTIL : modifier si on veut pouvoir modifier le nombre de recette affichée
+        $lengthAppartement = 5; // UTIL : modifier si on veut pouvoir modifier le nombre d'appartement affiché
         $_SESSION["appartementsPerPage"] = $lengthAppartement;
 
         if (array_key_exists("start", $_GET) && $_GET["start"] > 0) // si le paramettre de start n'est pas négatif
@@ -306,7 +306,7 @@ class AppartementController extends Controller {
     }
 
     /**
-     * permet d'ajouter ou de modifier sa note pour une recette
+     * permet d'ajouter ou de modifier sa note pour un appartement
      *
      * @return string
      */
@@ -327,7 +327,7 @@ class AppartementController extends Controller {
             $_GET["start"] = 0;
         }
 
-        if (array_key_exists("id", $_GET) && $database->AppartementExist($_GET["id"]))
+        if (array_key_exists("id", $_GET) && $database->AppartementExist($_GET["id"]) && $database->isRateExist($_SESSION["idUser"], $_GET["id"]))
         {
             $database->insertRating($_SESSION["idUser"], $_GET["id"]);
             $database->updateWish($_SESSION["idUser"], $_GET["id"], $database->GetVisitedStateForWish($_SESSION["idUser"], $_GET["id"]), 1);
@@ -353,7 +353,7 @@ class AppartementController extends Controller {
     }
 
     /**
-     * permet d'ajouter ou de modifier sa note pour une recette
+     * permet d'ajouter ou de modifier sa note pour un appartement
      *
      * @return string
      */
@@ -374,7 +374,7 @@ class AppartementController extends Controller {
             $_GET["start"] = 0;
         }
 
-        if (array_key_exists("id", $_GET) && $database->AppartementExist($_GET["id"]))
+        if (array_key_exists("id", $_GET) && $database->AppartementExist($_GET["id"]) && $database->isRateExist($_SESSION["idUser"], $_GET["id"]))
         {
             $database->removeRating($_SESSION["idUser"], $_GET["id"]);
             $database->updateWish($_SESSION["idUser"], $_GET["id"], $database->GetVisitedStateForWish($_SESSION["idUser"], $_GET["id"]), 0);
@@ -449,7 +449,7 @@ class AppartementController extends Controller {
     }
 
     /**
-     * permet d'ajouter une recette
+     * permet d'ajouter un appartement
      *
      * @return string
      */
@@ -529,9 +529,9 @@ class AppartementController extends Controller {
             $appartement["appImage"] = "defaultAppartementPicture.jpg";
             $appartement["appRate"] = 0;
 
-            $database->insertAppartement($appartement); // insertion de la recette dans la base de donnée
+            $database->insertAppartement($appartement); // insertion de l'appartement dans la base de donnée
 
-            $appartement = $database->getLastAppartement(); //get la recette pour la page edit (où l'on peut ajouter une image) 
+            $appartement = $database->getLastAppartement(); //get l'appartement pour la page edit (où l'on peut ajouter une image) 
 
             $view = file_get_contents('view/page/restrictedPages/manageAppartement/editAppartement.php');
         }
@@ -549,7 +549,7 @@ class AppartementController extends Controller {
     }
 
     /**
-     * permet de modifier une recette
+     * permet de modifier un appartement
      *
      * @return string
      */
@@ -689,7 +689,7 @@ class AppartementController extends Controller {
     }
 
     /**
-     * permet de supprimer une recette
+     * permet de supprimer un appartement
      *
      * @return string
      */
@@ -750,7 +750,7 @@ class AppartementController extends Controller {
     }
 
     /**
-     * permet de rendre l'index de départ rond par rapport au nombre de page et au nombre de recette par page <etc class=""></etc>
+     * permet de rendre l'index de départ rond par rapport au nombre de page et au nombre d'appartement par page <etc class=""></etc>
      *
      * @param int $startIndex 
      * @param [type] $database
@@ -783,11 +783,11 @@ class AppartementController extends Controller {
             }
             else if ($_GET["start"] == PHP_INT_MAX)
             {
-                if ($appartementNumber%$lengthAppartement == 0) // s'il y a pil le meme nombre de recette
+                if ($appartementNumber%$lengthAppartement == 0) // s'il y a pil le meme nombre d'appartement
                 {
                     $startIndex = $appartementNumber - ($lengthAppartement - $appartementNumber%$lengthAppartement);
                 }
-                else // s'il y a plus de recette
+                else // s'il y a plus d'appartement
                 {
                     $startIndex = $appartementNumber - $appartementNumber%$lengthAppartement;
                 }
